s/requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
Flask-Mail==0.9.1
Flask-Cors==3.0.10
python-dotenv==1.0.0
